Move C++ Val to its own header
Also bind `AnyRef` while I'm here with some small tests.
