+++ crates/c-api/include/wasmtime/types/val.hh
@@ -28,6 +28,8 @@ enum class ValKind {
   ExternRef,
   /// WebAssembly's `funcref` type from the reference types
   FuncRef,
+  /// WebAssembly's `anyref` type
+  AnyRef,
 };
 
 /// Helper X macro to construct statement for each enumerator in `ValKind`.
@@ -39,6 +41,7 @@ enum class ValKind {
   X(F64, "f64", WASM_F64)                                                      \
   X(ExternRef, "externref", WASM_EXTERNREF)                                    \
   X(FuncRef, "funcref", WASM_FUNCREF)                                          \
+  X(AnyRef, "anyref", WASMTIME_ANYREF)                                         \
   X(V128, "v128", WASMTIME_V128)
 
 /// \brief Used to print a ValKind.
